colorization-app/requirements.txt
# ─── Core ─────────────────────────────────────────────────────────────────────
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0
gunicorn>=21.2.0

# ─── Computer Vision ──────────────────────────────────────────────────────────
opencv-python>=4.9.0
opencv-contrib-python>=4.9.0
Pillow>=10.2.0
numpy>=1.26.0
scipy>=1.12.0

# ─── Deep Learning ────────────────────────────────────────────────────────────
torch>=2.2.0
torchvision>=0.17.0
# For GPU: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121

# ─── Real-ESRGAN ──────────────────────────────────────────────────────────────
realesrgan>=0.3.0
basicsr>=1.4.2

# ─── GFPGAN (optional) ────────────────────────────────────────────────────────
gfpgan>=1.3.8
facexlib>=0.3.0

# ─── Utilities ────────────────────────────────────────────────────────────────
tqdm>=4.66.0
requests>=2.31.0
python-dotenv>=1.0.0
